improved prediction and added reusable function
